move db test in own file
